Mention in README that hotel auto-starts on login
As a user, I was very surprised to discover this!  For example, it means that if you want your hotel server to have some environment variable set, `MY_VAR=... hotel start` won't work.  (Actually, on linux, it will work the first time, and then not subsequently.)  Anyways, it seems like a useful thing to document.  I even called it a feature ;-) .