Add an option to not run hotel on startup
This default is surely useful for many people, but it's not right for
me: I have many projects on my computer, and I only want to start hotel
if I'm working on certain ones.  Now there's an option to disable that
behavior.  I cribbed what it does from user-startup, minus the startup
parts.